% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todoist.R
\name{get_projects}
\alias{get_projects}
\title{Get projects}
\usage{
get_projects(
  .credentials = get_credentials_from_wrapper(wrapper),
  .perform = perform_by_default,
  .extract = extract_body_by_default,
  .extractor = extractor
)
}
\arguments{
\item{...}{Arguments passed to the requestor function}

\item{credentials}{Credential}

\item{action}{dryrun or perform}

\item{decode_if_success}{Decode if success?}
}
\value{
a list of projects
}
\description{
Get a list of projects
}
